Split out stable schema code and make constants global
